Move paletted raster renderer tests to own file
