+++ tests/src/python/CMakeLists.txt
@@ -200,6 +200,7 @@ ADD_PYTHON_TEST(PyQgsOptional test_qgsoptional.py)
 ADD_PYTHON_TEST(PyQgsOrientedBox3D test_qgsorientedbox3d.py)
 ADD_PYTHON_TEST(PyQgsOwsConnection test_qgsowsconnection.py)
 ADD_PYTHON_TEST(PyQgsPainting test_qgspainting.py)
+ADD_PYTHON_TEST(PyQgsPalettedRasterRenderer test_qgspalettedrasterrenderer.py)
 ADD_PYTHON_TEST(PyQgsPalLabelingBase test_qgspallabeling_base.py)
 ADD_PYTHON_TEST(PyQgsPalLabelingCanvas test_qgspallabeling_canvas.py)
 ADD_PYTHON_TEST(PyQgsPalLabelingLayout test_qgspallabeling_layout.py)
